Django-Middleware-0x03/requirement.txt
asgiref==3.8.1
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.2
